# reorganize code, little bit faster