@gunnarmorling 2024-01-21
Nice, 00:02.336 now!

```
Benchmark 1: timeout -v 300 ./calculate_average_abeobk.sh 2>&1
  Time (mean ± σ):      2.336 s ±  0.006 s    [User: 15.820 s, System: 0.701 s]
  Range (min … max):    2.329 s …  2.343 s    5 runs

Summary
  abeobk: trimmed mean 2.3362541187466674, raw times 2.3407769550800004,2.32911485408,2.3351161250800003,2.3429942460800004,2.3328692760800003

Leaderboard

| # | Result (m:s.ms) | Implementation     | JDK | Submitter     | Notes     |
|---|-----------------|--------------------|-----|---------------|-----------|
|   | 00:02.336 | [link](https://github.com/gunnarmorling/1brc/blob/main/src/main/java/dev/morling/onebrc/CalculateAverage_abeobk.java)| 21.0.1-graal | [Van Phu DO](https://github.com/abeobk) | GraalVM native binary |
```